Created tables, added in items

+++ code/Models/item.cs
@@ -17,7 +17,7 @@ namespace code.Models
         [Key]
         public int itemID { get; set; }
         public string itemName { get; set; }
-        public float itemPrice { get; set; }
+        public double itemPrice { get; set; }
         public string imgPath { get; set; }
         public string itemDescription { get; set; }
